dist/datachannel_test_2peer/send/007.html

<!DOCTYPE html>
<html lang="en" class="a">
	<head>
		<meta charset="utf-8">
		<!-- Testharness Timeout to long -> 60 Seconds-->
		<meta name="timeout" content="long">
		<title>WebRTC - Send Test</title>
		<script src="../resources/testharness.js"></script>
		<script src="../resources/testharnessreport.js"></script><link href="../resources/testharness.css" type="text/css" rel="stylesheet"/>
		<!-- Including the Firebase.js to send the ICE Candidate to the other peer-->
		<script type="text/javascript" src='https://cdn.firebase.com/v0/firebase.js'></script>
	</head>
	<body class="a">
		<div>
			<div id="signallingInfo"></div>
			
				
			
		</div>
		<hr />

		<!-- Testharness log -->
		<div id="log"></div>
		<script src="../datachannel2peers.js" type="text/javascript"></script>
		<script type="text/javascript">
            var label = "test";
            var dataChannel = null;
            var expected = "arraybuffer";
			var buffer = new ArrayBuffer(12);
	        var x = new Int32Array(buffer);
	        x[0] = 11111111;
	        x[1] = 22222222;
	        x[2] = 33333333;
            var test = async_test("Set up a DataChannel - send an ArrayBuffer (Int32Array) and receive the data as an ArrayBuffer - check type and data");
            if (offerer) {
                test.step(function() {
                    dataChannel = peerConnection.createDataChannel(label);
                    dataChannel.onopen = test.step_func(function() {
                    	try{
                    		dataChannel.send(x.buffer);
                    	}catch(e){
                    		assert_unreached(e.name + ": " + e.message);
                    	}
                        test.done();
                    });
                });
            } else {
                peerConnection.ondatachannel = test.step_func(function(e) {
                    dataChannel = e.channel;
                    try {
                        dataChannel.binaryType = expected;
                    } catch(e) {
                        //assert_unreached(e.name + ": " +e.message);
                    }
                    dataChannel.onmessage = test.step_func(function(e) {
						var slice = e.data.slice();
			            var y = new Int32Array(slice);
			            assert_array_equals(y, x, "Data not send correct ");
			            assert_true(e.data instanceof ArrayBuffer, "No blob received ");
			            assert_equals(dataChannel.binaryType, expected, "Wrong binaryType: ");
			            test.done();
                    });
                });
            }
		</script>
	</body>
</html>

